Path for tokens fix and 6 dig code instead 4

+++ Application/Dto/CodeVerify.cs
@@ -2,6 +2,6 @@
 {
     public class CodeVerify
     {
-        public short Code { get; set; }
+        public int Code { get; set; }
     }
 }
